requirements.txt
mistralai
requests
numpy
faiss-cpu 